Shell script tooling update

+++ srcs/requirements/tools/cleanup.sh
@@ -11,16 +11,21 @@ if [ -d "$SECRETS" ]; then
   rm -rf "$SECRETS"
 fi
 
-if docker volume ls | grep -o 'inception_[a-z-]\+'; then
+if docker volume ls | grep -oq 'inception_[a-z-]\+'; then
   docker volume rm $(docker volume ls | grep -o 'inception_[a-z-]\+')
 fi
+
 VOLUME="$(grep VOLUME $ENV | cut -d= -f2)"
 if [ -d "$VOLUME" ]; then
-  sudo rm -rf "$VOLUME"
+  read -p "Do you want to permanently delete your volume data? [y/N] " RESULT
+  if [ -n "$RESULT" ] && [ $RESULT == "y" ] ; then
+    sudo rm -rf "$VOLUME"
+    echo "Directory $VOLUME was deleted!"
+  fi
 fi
 
 DOMAIN="$(grep DOMAIN $ENV | cut -d= -f2)"
-if grep "127.0.0.42 $DOMAIN" /etc/hosts; then
+if grep -q "127.0.0.42 $DOMAIN" /etc/hosts; then
   sudo sed -i "/127.0.0.42 $DOMAIN/d" /etc/hosts
 fi
